refactor: remove dto and subscriber folder
