--- packages/deployment-service/src/dto/index.ts
@@ -1,2 +1,0 @@
-export * from './pipeline-dto'
-export * from './parameter-dto'
